Roles, permissions, and seeders

+++ app/Http/Controllers/Api/PostController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\StorePostRequest;
 use App\Http\Resources\PostResource;
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
@@ -50,6 +51,8 @@ class PostController extends Controller
 
     public function store(StorePostRequest $request)
     {
+        Gate::authorize('posts.create');
+
         // We're just going to log file uploads, not actually do anything with them...
         if($request->hasFile('thumbnail')) {
             $filename = $request->file('thumbnail')->getClientOriginalName();
@@ -63,11 +66,13 @@ class PostController extends Controller
 
     public function show(Post $post)
     {
+        Gate:: authorize(('posts.update'));
         return new PostResource($post);
     }
 
     public function update(Post $post, StorePostRequest $request) 
     {
+        Gate::authorize('posts.update');
         $post->update($request->validated());
 
         return new PostResource($post);
@@ -75,6 +80,7 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+        Gate::authorize('posts.delete');
         $post->delete();
 
         return response()->noContent();
